Move rule engine into separate package
