+++ src/main/java/net/projecteuler/util/numbertoword/Number.java
@@ -1,0 +1,4 @@
+package net.projecteuler.util.numbertoword;
+
+public record Number(int number, String wordRepresentation) {
+}
