+++ src/main/java/net/projecteuler/util/numbertoword/NumberResult.java
@@ -1,0 +1,4 @@
+package net.projecteuler.util.numbertoword;
+
+public record NumberResult(String value) {
+}
