+++ src/main/java/net/projecteuler/util/numbertoword/NumberRule.java
@@ -1,0 +1,7 @@
+package net.projecteuler.util.numbertoword;
+
+import java.util.Optional;
+
+public interface NumberRule {
+  Optional<NumberResult> evaluate(int number);
+}
