+++ src/main/java/net/projecteuler/util/numbertoword/NumberToWordRuleEngine.java
@@ -1,0 +1,27 @@
+package net.projecteuler.util.numbertoword;
+
+import net.projecteuler.util.numbertoword.rules.DoubleDigitNumberFromTwentyUpToHundredRule;
+import net.projecteuler.util.numbertoword.rules.DoubleDigitNumberUpToTwentyRule;
+import net.projecteuler.util.numbertoword.rules.OneThousandNumberRule;
+import net.projecteuler.util.numbertoword.rules.SingleDigitNumberRule;
+import net.projecteuler.util.numbertoword.rules.ThreeDigitNumberRule;
+
+import java.util.List;
+import java.util.Optional;
+
+public class NumberToWordRuleEngine {
+  private final List<NumberRule> rules = List.of(
+    new SingleDigitNumberRule(),
+    new DoubleDigitNumberUpToTwentyRule(),
+    new DoubleDigitNumberFromTwentyUpToHundredRule(),
+    new ThreeDigitNumberRule(),
+    new OneThousandNumberRule()
+  );
+
+  public NumberResult process(int number) {
+    return rules.stream()
+      .map(rule -> rule.evaluate(number))
+      .flatMap(Optional::stream)
+      .findFirst().orElseThrow();
+  }
+}
